beginning to refactor octree code to be less OOP
now most logic happens in octree class and not the octNode or OctValue class. This will help massively, because an OOP doesn't work when the element you're operating on are structs in arrays, because in c# you can't even get a ref to a struct in an array (in safe code land) so methods which act upon the object itself don't make sense as you have to then copy yourself into an array for the state to be persisted at the end of every method... I could have went into unsafe c# land and avoid the perf overhead of copying struct values, but I don't think that's a REAL perf cost, and maybe I can still do that later for practice.